first attempt to fold global config into config schema
