+++ cylc/flow/glbl_cfg_lazy.py
@@ -1,0 +1,22 @@
+# THIS FILE IS PART OF THE CYLC SUITE ENGINE.
+# Copyright (C) 2008-2019 NIWA & British Crown (Met Office) & Contributors.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""Allow lazy loading of `cylc.flow.cfgspec.globalcfg`."""
+
+
+def glbl_cfg(cached=True):
+    """Load and return the global configuration singleton instance."""
+    from cylc.flow.config_schema import GlobalConfig
+    return GlobalConfig.get_inst(cached=cached)
